Correctly recognize disconnected devices (#2)
If pselect is 1, this means one can call read on the file descriptor (normally this indicates data is available).
However, if then read() returns zero bytes, this indicates the device was disconnected.
https://stackoverflow.com/questions/34170350/detecting-if-a-character-device-has-disconnected-in-linux-in-with-termios-api-c

In this case, throw an exception.
Also, slightly improve error handling if device or sensor is not found.